app windows/release-proxyfix/win-unpacked/resources/bridge-runtime/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-jose>=3.3.0
passlib>=1.7.4
openai==1.6.1
httpx>=0.23.0,<0.28
python-multipart>=0.0.9
protobuf>=4.21.0
cryptography>=41.0.0
websockets>=12.0
faster-whisper>=1.1.1
